[Ref] mpt/io_read/filecursor.hpp: Prepare for 64bit pos_type > size_t. See <https://bugs.openmpt.org/view.php?id=1647>.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@19336 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ src/mpt/io_read/filecursor.hpp
@@ -252,7 +252,7 @@ public:
 			size_ = 0;
 			pinnedData = nullptr;
 			if (!file.CanRead(size)) {
-				size = file.BytesLeft();
+				size = static_cast<std::size_t>(file.BytesLeft());
 			}
 			size_ = size;
 			if (file.DataContainer().HasPinnedView()) {
